+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/ColumnStorageFacade.java
@@ -29,7 +29,7 @@ public final class ColumnStorageFacade<S, T> implements ColumnStorage<T> {
   }
 
   @Override
-  public int typeSize() {
+  public long typeSize() {
     throw new UnsupportedOperationException("Not implemented");
   }
